fix var assign but errors at arrays

+++ include/arrcopy.h
@@ -3,18 +3,19 @@
 #define COPY_H
 
 #include <stdbool.h>
+#include <stdlib.h>
 
 /* Copy passed array of integers into a new one */
-int *intcopy(int *src, int num_elems);
+int *intcopy(int *src, size_t num_elems);
 
 /* Copy passed array of bools into a new one */
-bool *boolcopy(bool *src, int num_elems);
+bool *boolcopy(bool *src, size_t num_elems);
 
 /* Copy passed array of floats into a new one */
-float *floatcopy(float *src, int num_elems);
+float *floatcopy(float *src, size_t num_elems);
 
 /* Copy passed array of doubles into a new one */
-double *doublecopy(double *src, int num_elems);
+double *doublecopy(double *src, size_t num_elems);
 
 /* Copy passed string into a new one */
 char *strcopy(char *src);
